pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "course-rag-chatbot"
version = "0.1.0"
description = "MVP RAG chatbot for university course transcripts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "streamlit>=1.38.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "numpy>=1.26.0",
  "faiss-cpu>=1.8.0",
  "openai>=1.40.0",
  "cohere>=5.8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]
transcription = [
  "faster-whisper>=1.0.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]
